Started restructuring work (to make ZeGrapher usable as a lib)
The goal of this code reorganisation is to stick to a model/view
philosophy and make ZeGrapher usable as a library in other projects.
This will also make the source code more elegant and structured.
We'll adopt Qt's approach in naming by prefixing all ZeGrapher's classes
with "Ze"

Work started:
- Merged classes TreeCreator and ExprCalcultor into ZeExpression. This
class will be the base of all math calculations. Started modifiying the
code.